[BLAS] Remove queue from `host_task_internal` API
A `sycl::queue` object is passed to the `host_task_internal` function
as a parameter and used when interacting with a object
`CublasScopedContextHandler`. However, the `interop_handle` parameter
that is also passed to this entry-point and `CublasScopedContextHandler`
constructor is already able to get this information via
`interop_handle::get_native_queue()`.

This patch removes the superfluous queue parameters and changes the
`CublasScopedContextHandler:get_stream()` implementation to use
the `interop_handle` method instead. The unused `get_context()` method
is also removed.