[BLAS] Avoid sync on native-command
The `CUBLAS_ERROR_FUNC_SYNC` macro calls `cuStreamSynchronize(currentStreamId);`
to synchronize the stream associated with the cuBLAS handle after making
the API call, which `CUBLAS_ERROR_FUNC` doesn't.

When using the
[sycl_ext_codeplay_enqueue_native_command](https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/experimental/sycl_ext_codeplay_enqueue_native_command.asciidoc)
extension is advised by the specification not to use commands that block

> The function object interopCallable is invoked to enqueue commands to a native queue or graph and therefore, APIs which block or synchronize could prolong or interfere with other commands being enqueued to the backend.

As these will run when the command is submitted before the command-group
dependencies have been satisfied.

Instead use the sync variant for the host-task path rather than native
enqueue path, which aligns with how
[cublas_batch.cpp](https://github.com/uxlfoundation/oneMath/blob/f09d6cbe2aed9ffef47a6b36b7598322e7cb3104/src/blas/backends/cublas/cublas_batch.cpp#L170)
is using a sync variant on the host-task path and not on the native
enqueue path.